+++ mobile/android/android-components/components/feature/qr/src/main/java/mozilla/components/feature/qr/views/AutoFitTextureView.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License. */
 
-package mozilla.components.support.base.android.view
+package mozilla.components.feature.qr.views
 
 import android.content.Context
 import android.util.AttributeSet